+++ src/CoordinatorClient.h
@@ -41,7 +41,7 @@ class CoordinatorClient : public Client {
 
     void createTable(const char* name, uint32_t serverSpan = 1);
     void dropTable(const char* name);
-    uint32_t openTable(const char* name);
+    uint64_t openTable(const char* name);
 
     ServerId enlistServer(ServerId replacesId,
                           ServiceMask serviceMask,
